pyproject.toml
[tool.poetry]
name = "em-fast-api-project"
version = "0.1.0"
description = ""
authors = ["Maxim Olesov <windigotp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
pydantic-settings = "^2.3.4"
asyncpg = "0.29.0"
pyjwt = {version = "2.8.0", extras = ["crypto"]}
bcrypt = "4.2.0"
orjson = "3.10.6"
redis = "5.0.7"
celery = "5.4.0"
flower = "2.0.1"


[tool.poetry.group.test.dependencies]
pytest = "8.2.2"
pytest-dotenv = "0.5.2"
pytest-asyncio = "0.23.7"


[tool.poetry.group.dev.dependencies]
ruff = "0.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
